Refactor the notification options into a separate component.
